Rename Context to ContextObject, Create type for JsonLDContext
