+++ ActivityPub.Types/Json/JsonLDContextConverter.cs
@@ -8,10 +8,14 @@ using ActivityPub.Types.Util;
 namespace ActivityPub.Types.Json;
 
 /// <summary>
-/// Custom converter for <see cref="JsonLDContext"/>
+/// Custom converter for the JSON-LD "@context" property.
 /// </summary>
+/// <remarks>
+/// We need THREE FUCKING CONVERTERS for a minimum-viable implementation!
+/// </remarks>
 /// <seealso cref="JsonLDTermConverter"/>
-/// <seealso cref="JsonLDContextPropertyConverter"/>
+/// <seealso cref="JsonLDContextObjectConverter"/>
+/// <seealso href="https://www.w3.org/TR/json-ld11/#the-context"/>
 public class JsonLDContextConverter : JsonConverter<JsonLDContext>
 {
     public override JsonLDContext? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
@@ -22,34 +26,57 @@ public class JsonLDContextConverter : JsonConverter<JsonLDContext>
                 return null;
 
             case JsonTokenType.String:
-            {
-                var str = reader.GetString()!;
-                return new JsonLDContext(str);
-            }
-
             case JsonTokenType.StartObject:
             {
-                var terms = JsonSerializer.Deserialize<Dictionary<string, JsonLDTerm>>(ref reader, options)?.AsReadOnly();
-                if (terms == null)
-                    throw new JsonException("Failed to parse JsonLDContext terms");
+                var context = ReadContext(ref reader, options);
+                return new JsonLDContext(new HashSet<JsonLDContextObject>
+                {
+                    context
+                } );
+            }
 
-                return new JsonLDContext(terms);
+            case JsonTokenType.StartArray:
+            {
+                var set = new HashSet<JsonLDContextObject>();
+                while (reader.Read() && reader.TokenType != JsonTokenType.EndArray)
+                {
+                    var context = ReadContext(ref reader, options);
+                    set.Add(context);
+                }
+
+                return new JsonLDContext(set);
             }
 
             default:
-                throw new JsonException($"Cannot deserialize {reader.TokenType} as JsonLDContext");
+                throw new JsonException($"Cannot deserialize {reader.TokenType} as @context field");
         }
+    }
+
+    private static JsonLDContextObject ReadContext(ref Utf8JsonReader reader, JsonSerializerOptions options)
+    {
+        var context = JsonSerializer.Deserialize<JsonLDContextObject>(ref reader, options);
+        if (context == null)
+            throw new JsonException($"Failed to parse object-form @context field");
+
+        return context;
     }
 
     public override void Write(Utf8JsonWriter writer, JsonLDContext value, JsonSerializerOptions options)
     {
-        if (value.IsExternal)
+        if (value.ContextObjects.Count == 1)
         {
-            writer.WriteStringValue(value.ExternalLink);
+            var context = value.ContextObjects.First();
+            JsonSerializer.Serialize(writer, context, options);
         }
         else
         {
-            JsonSerializer.Serialize(writer, value.Terms, options);
+            writer.WriteStartArray();
+            foreach (var context in value.ContextObjects)
+            {
+                JsonSerializer.Serialize(writer, context, options);
+            }
+
+            writer.WriteEndArray();
         }
     }
 }
